Fix CI regressions in service auth and trade manager stubs
