+++ bot/trade_manager/__init__.py
@@ -6,16 +6,26 @@
 ``sys.path`` или ``noqa``.
 """
 
+from __future__ import annotations
+
+import importlib
+import sys
+from typing import Any
+
 from bot.config import OFFLINE_MODE
 
 if OFFLINE_MODE:
     from services.offline import OfflineBybit as TradeManager
-    from services.offline import OfflineTelegram as TelegramLogger
+    from services.offline import OfflineTelegram as _OfflineTelegram
 
     __all__ = ["TradeManager", "TelegramLogger"]
+
+    def __getattr__(name: str) -> Any:  # pragma: no cover - simple passthrough
+        if name == "TelegramLogger":
+            return _OfflineTelegram
+        raise AttributeError(name)
 else:  # pragma: no cover - реальная инициализация
     from bot.http_client import close_async_http_client, get_async_http_client
-    from bot.utils import TelegramLogger
 
     from .core import TradeManager
     from .service import (
@@ -47,3 +57,11 @@ else:  # pragma: no cover - реальная инициализация
         "get_http_client",
         "close_http_client",
     ]
+
+    def __getattr__(name: str) -> Any:
+        if name == "TelegramLogger":
+            utils = sys.modules.get("utils")
+            if utils is None:
+                utils = importlib.import_module("utils")
+            return utils.TelegramLogger
+        raise AttributeError(name)
