+++ services/trade_manager_service.py
@@ -23,7 +23,7 @@ except Exception:  # pragma: no cover - fallback when flask.typing missing
     ResponseReturnValue = Any  # type: ignore
 
 from bot.trade_manager import server_common
-from bot.utils import validate_host, safe_int
+from bot.shared_utils import validate_host, safe_int
 from services.logging_utils import sanitize_log_value
 
 server_common.load_environment()
@@ -36,13 +36,23 @@ logger = logging.getLogger(__name__)
 POSITIONS_LOCK = threading.RLock()
 
 _exchange_runtime: server_common.ExchangeRuntime | None = None
+exchange: Any | None = None
 
 
 def _current_exchange() -> Any | None:
+    """Return the currently bound exchange instance if available."""
+
+    global exchange
+
     runtime = _exchange_runtime
     if runtime is None:
-        return None
-    return runtime.current()
+        return exchange
+
+    current = runtime.current()
+    if current is not None:
+        exchange = current
+        return current
+    return exchange
 
 
 def init_exchange() -> None:
@@ -51,7 +61,8 @@ def init_exchange() -> None:
     runtime = _exchange_runtime
     if runtime is None:
         raise RuntimeError("Exchange runtime is not configured")
-    runtime.init()
+    global exchange
+    exchange = runtime.init()
 
 
 # Expected API token for simple authentication
@@ -66,7 +77,10 @@ if hasattr(app, "before_first_request"):
 def _bind_exchange() -> None:
     runtime = _exchange_runtime
     if runtime is not None:
-        runtime.bind()
+        bound = runtime.bind()
+        if bound is not None:
+            global exchange
+            exchange = bound
 
 
 @app.before_request
@@ -344,6 +358,7 @@ _exchange_runtime = server_common.ExchangeRuntime(
     after_create=_sync_positions_with_exchange,
 )
 exchange_provider = _exchange_runtime.provider
+ccxt = _exchange_runtime.ccxt
 CCXT_BASE_ERROR = _exchange_runtime.ccxt_base_error
 CCXT_NETWORK_ERROR = _exchange_runtime.ccxt_network_error
 CCXT_BAD_REQUEST = _exchange_runtime.ccxt_bad_request
