+++ trade_manager.py
@@ -5,6 +5,7 @@
 
 from __future__ import annotations
 
+import importlib
 import warnings
 
 warnings.warn(
@@ -12,5 +13,10 @@ warnings.warn(
     DeprecationWarning,
     stacklevel=2,
 )
+
+_tm = importlib.import_module("bot.trade_manager")
+# Reload the package so that dynamic stubs injected into ``bot.utils`` during
+# tests are respected even when ``bot.trade_manager`` was previously imported.
+_tm = importlib.reload(_tm)
 
 from bot.trade_manager import *  # noqa: F401,F403
